refactor(cmd): extract CLI mode logic to cli.go
